Functions to interact with modLand

ModLand is an online archive containing over 400,000 module files. These functions allow you to search in and download from this archive.

modland_search(text, ...)

modland_download(
  format,
  author,
  title,
  mirror = c("modland.com", "ftp.modland.com", "antarctica.no", "ziphoid.com",
    "exotica.org.uk"),
  read_fun = read_mod,
  ...
)

Arguments

text

Search text, to look for on modland.

...

Arguments passed on to read_fun

format

A single length character vector, indicating the tracker file format. Can be obtained from a modland_search().

author

A single length character vector, indicating the module author name. Can be obtained from a modland_search().

title

A single length character vector, indicating the module title. Can be obtained from a modland_search().

mirror

A single length character vector. Should contain one of the mirrors listed in the 'usage' section. Select a mirror site from which the module file needs to be downloaded.

read_fun

Function that accepts an URL first argument. By default it is read_mod() and is used to read the file. You can replace it with other functions such as ProTrackR2::pt2_read_mod().

Value

In case of modland_search() a data.frame with search results are returned (or NULL if there are no results).

modland_download() will return the result of the function specified by read_fun. By default it will return an openmpt class object.

See also

Examples

search_result <- modland_search("elekfunk mod")

## The URL in the search results will download a rendered
## ogg file. If you want to download te original mod file,
## use this: 
if (length(search_result) > 0) {
  mod <- modland_download(search_result$format[[1]],
                          search_result$author[[1]],
                          search_result$title[[1]])
}