openmpt: Open ModPlug Tracker Port

Tracker music uses audio samples and pattern tables to organise musical compositions. Such music is stored in module files. This package reads, renders and plays module files using the 'libopenmpt' library https://lib.openmpt.org/.

See also

Author

Maintainer: Pepijn de Vries pepijn.devries@outlook.com (ORCID)

Other contributors:

Jeroen Ooms jeroenooms@gmail.com (ORCID) [contributor]

Jester (Copyright holder of cyberrid.mod) [copyright holder, data contributor]

OpenMPT Project Developers and Contributors [copyright holder]